[8주차] 마크/한재민

+++ umc 8주차 과제/demo/src/main/java/umc/study/repository/UserRepository/UserRepository.java
@@ -3,5 +3,5 @@ package umc.study.repository.UserRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import umc.study.domain.User;
 
-public interface UserRepository extends JpaRepository<User, Integer>,UserRepositoryCustom {
+public interface UserRepository extends JpaRepository<User, Long>,UserRepositoryCustom {
 }
